extract ini config to separate service
