+++ src/PrestaShopBundle/Controller/Admin/Configure/AdvancedParameters/ImportController.php
@@ -42,7 +42,7 @@ use Symfony\Component\HttpFoundation\ResponseHeaderBag;
 class ImportController extends FrameworkBundleAdminController
 {
     /**
-     * Show import form
+     * Show import form & handle forwarding to legacy controller
      *
      * @param Request $request
      *
@@ -91,7 +91,7 @@ class ImportController extends FrameworkBundleAdminController
             'file_upload_url' => $this->generateUrl('admin_import_file_upload'),
             'import_file_names' => $names,
             'import_dir' => $this->get('prestashop.core.import.dir')->getDir(),
-            'max_file_upload_size' => $this->getPostMaxSizeInBytes(),
+            'max_file_upload_size' => $this->get('prestashop.utils.ini_config')->getPostMaxSizeInBytes(),
         ];
     }
 
@@ -184,34 +184,5 @@ class ImportController extends FrameworkBundleAdminController
         $legacyImportUrl = $legacyContext->getAdminLink($legacyController);
 
         return $this->redirect($legacyImportUrl, Response::HTTP_TEMPORARY_REDIRECT);
-    }
-
-    /**
-     * Get m post max size from ini configuration in bytes
-     *
-     * @return int
-     */
-    private function getPostMaxSizeInBytes()
-    {
-        $post_max_size = ini_get('post_max_size');
-        $bytes = (int) trim($post_max_size);
-        $last = strtolower($post_max_size[strlen($post_max_size) - 1]);
-
-        switch ($last) {
-            case 'g':
-                $bytes *= 1024;
-                // no break to fall-through
-            case 'm':
-                $bytes *= 1024;
-                // no break to fall-through
-            case 'k':
-                $bytes *= 1024;
-        }
-
-        if (!isset($bytes) || $bytes == '') {
-            $bytes = 20971520;  // 20Mb
-        }
-
-        return $bytes;
     }
 }
