+++ src/PrestaShopBundle/Resources/config/services/services.yml
@@ -174,6 +174,9 @@ services:
     prestashop.utils.zip_manager:
         class: PrestaShopBundle\Utils\ZipManager
 
+    prestashop.utils.ini_config:
+        class: PrestaShopBundle\Utils\IniConfig
+
     #TRANSLATIONS FACTORIES
     prestashop.translation.translations_factory:
         class: PrestaShopBundle\Translation\Factory\TranslationsFactory
